Removed DatafloRunner, mimicking Flask execution api
